requirements.txt
flask
jinja2
opencv-contrib-python
mysql-connector-python
numpy
pandas
werkzeug